+++ ltt/branches/poly/ltt/marker.h
@@ -66,8 +66,8 @@ static inline struct marker_info *marker_get_info_from_name(LttTrace *trace,
   gpointer orig_key, value;
   int res;
 
-  res = g_hash_table_lookup_extended(trace->markers_hash, (gconstpointer)name,
-    &orig_key, &value);
+  res = g_hash_table_lookup_extended(trace->markers_hash,
+    (gconstpointer)(gulong)name, &orig_key, &value);
   if (!res)
     return NULL;
   return marker_get_info_from_id(trace, (guint16)(gulong)value);
@@ -80,7 +80,7 @@ static inline char *marker_get_format_from_name(LttTrace *trace,
   int res;
 
   res = g_hash_table_lookup_extended(trace->markers_format_hash,
- 	 	(gconstpointer)name, &orig_key, &value);
+ 	 	(gconstpointer)(gulong)name, &orig_key, &value);
   if (!res)
     return NULL;
   return (char *)value;
